# docs/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LogicPOS.ApiServer | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/worksessions/periods/open-day": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/periods/open-day": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/periods/open-day": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/periods/open-day": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksessions/periods/close-day": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/periods/close-day": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/periods/close-day": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/periods/close-day": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseWorkSessionDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/worksessions/periods/terminal-is-open": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/periods/terminal-is-open": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/periods/terminal-is-open": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/periods/terminal-is-open": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksessions/periods/open-session": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/periods/open-session": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/periods/open-session": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/periods/open-session": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksessions/periods/close-session": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/periods/close-session": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/periods/close-session": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/periods/close-session": {
      "put": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseTerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/worksessions/period/terminal/{terminalId}": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/period/terminal/{terminalId}": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/period/terminal/{terminalId}": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/period/terminal/{terminalId}": {
      "get": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksessions/movements/cash-drawer-in": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/movements/cash-drawer-in": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/movements/cash-drawer-in": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksession/movements/cash-drawer-in": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksessions/movements/cash-drawer-out": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worksession/movements/cash-drawer-out": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/worksessions/movements/cash-drawer-out": {
      "post": {
        "tags": [
          "WorkSessions"
        ],
        "parameters": [
          {
            "name": "terminalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CashDrawerMovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashDrawerMovementRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CloseTerminalSessionRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CloseWorkSessionDayRequest": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OpenTerminalSessionRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OpenWorkSessionDayRequest": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}